Update collection header
