+++ frontend/src/components/VCollectionHeader/meta/VCollectionHeader.stories.mdx
@@ -45,7 +45,7 @@ export const imageProviderNames = [
 
 export const AllCollectionsTemplate = (args) => ({
   template: `
-<div class="wrapper w-full p-2 flex flex-col gap-2 bg-dark-charcoal-06">
+<div class="wrapper w-full p-3 flex flex-col gap-4 bg-dark-charcoal-06">
     <VCollectionHeader v-for="collection in args.collections" :key="collection.collectionName" v-bind="collection" class="bg-white"/>
 </div>`,
   components: { VCollectionHeader },
