+++ frontend/test/locales/ar.json
@@ -701,6 +701,31 @@
     "none": "لا توجد عمليات بحث حديثة لعرضها.",
     "disclaimer": "لا يقوم Openverse بتخزين عمليات البحث الأخيرة ، يتم الاحتفاظ بهذه المعلومات محليًا في متصفحك."
   },
+  "collection": {
+    "heading": {
+      "tag": "الكلمات الدلالية",
+      "creator": "مبتدع",
+      "source": "منشأ"
+    },
+    "link": {
+      "source": "موقع مفتوح المصدر",
+      "creator": "افتح صفحة المنشئ"
+    },
+    "ariaLabel": {
+      "creator": {
+        "audio": "الملفات الصوتية بواسطة {creator}",
+        "image": "الصور عن طريق {creator}"
+      },
+      "source": {
+        "audio": "ملفات صوتية من {source}",
+        "image": "الصور عن طريق {source}"
+      },
+      "tag": {
+        "audio": "الملفات الصوتية مع العلامة {tag}",
+        "image": "الصور التي تحمل الوسم {tag}"
+      }
+    }
+  },
   "unknownError": {
     "heading": "عفوًا ، يبدو أنه حدث خطأ ما. حاول مرة اخرى."
   }
